final tunes

+++ book/chapter-5/5.7-make-glibc.sh
@@ -10,26 +10,35 @@ echo "Required disk space: 823 MB"
 # matching, arithmetic, and so on
 tar -xf /sources/glibc-*.tar.xz -C /tmp/ \
   && mv /tmp/glibc-* /tmp/glibc \
-  && pushd /tmp/glibc \
-  && mkdir -v build \
-  && cd build \
-  && ../configure                       \
-    --prefix=/tools                    \
-    --host=$LFS_TGT                    \
-    --build=$(../scripts/config.guess) \
-    --enable-kernel=3.2                \
-    --with-headers=/tools/include      \
-    libc_cv_forced_unwind=yes          \
-    libc_cv_c_cleanup=yes              \
-  && make \
-  && make install \
-  && popd \
+  && pushd /tmp/glibc
+
+mkdir -v build
+cd       build
+
+../configure                             \
+      --prefix=/tools                    \
+      --host=$LFS_TGT                    \
+      --build=$(../scripts/config.guess) \
+      --enable-kernel=3.2             \
+      --with-headers=/tools/include      \
+      libc_cv_forced_unwind=yes          \
+      libc_cv_c_cleanup=yes
+
+# Compile the package:
+make
+
+# Install the package:
+make install
+
+popd \
   && rm -rf /tmp/glibc
 
 # perform a sanity check that basic functions (compiling and linking)
 # are working as expected
-cd /sources
-echo 'int main(){}' > dummy.c \
-  && $LFS_TGT-gcc dummy.c \
-  && readelf -l a.out | grep ': /tools' \
-  && rm -v dummy.c a.out
+if [ $LFS_TEST -eq 1 ]; then
+    # perform a sanity check
+    echo 'int main(){}' > dummy.c \
+      && $LFS_TGT-gcc dummy.c \
+      && readelf -l a.out | grep ': /tools' \
+      && rm -v dummy.c a.out
+fi
